+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
+- Files not being able to be re-added to files and camera element after being removed
 - inconsistent margin for form element error messages
 
 ## 0.1.11 (2021-02-15)